+++ test/functional/android/android/device_test.rb
@@ -187,7 +187,7 @@ class AppiumLibCoreTest
         rect = el.rect
 
         Appium::Core::TouchAction.new(@@driver)
-                                 .swipe(start_x: 75, start_y: 500, offset_x: 75, offset_y: 500, duration: 500)
+                                 .swipe(start_x: 75, start_y: 500, end_x: 75, end_y: 500, duration: 500)
                                  .perform
         @@driver.back # The above command become "tap" action since it doesn't move.
         el = @@core.wait { @@driver.find_element :accessibility_id, 'Fragment' }
@@ -195,7 +195,7 @@ class AppiumLibCoreTest
         assert rect.y == el.rect.y
 
         touch_action = Appium::Core::TouchAction.new(@@driver)
-                                                .swipe(start_x: 75, start_y: 500, offset_x: 75, offset_y: 300, duration: 500)
+                                                .swipe(start_x: 75, start_y: 500, end_x: 75, end_y: 300, duration: 500)
 
         assert_equal :press, touch_action.actions[0][:action]
         assert_equal({ x: 75, y: 500 }, touch_action.actions[0][:options])